Separate Generator code by schema type.
