+++ tests/schemas/test_openapi.py
@@ -5,8 +5,7 @@ from django.test import RequestFactory, TestCase, override_settings
 from rest_framework import filters, generics, pagination, serializers
 from rest_framework.compat import uritemplate
 from rest_framework.request import Request
-from rest_framework.schemas.generators import OpenAPISchemaGenerator
-from rest_framework.schemas.openapi import AutoSchema
+from rest_framework.schemas.openapi import AutoSchema, SchemaGenerator
 
 from . import views
 
@@ -18,7 +17,7 @@ def create_request(path):
 
 
 def create_view(view_cls, method, request):
-    generator = OpenAPISchemaGenerator()
+    generator = SchemaGenerator()
     view = generator.create_view(view_cls.as_view(), method, request)
     return view
 
@@ -144,7 +143,7 @@ class TestGenerator(TestCase):
         patterns = [
             url(r'^example/?$', views.ExampleListView.as_view()),
         ]
-        generator = OpenAPISchemaGenerator(patterns=patterns)
+        generator = SchemaGenerator(patterns=patterns)
         generator._initialise_endpoints()
 
         paths = generator.get_paths()
@@ -160,7 +159,7 @@ class TestGenerator(TestCase):
         patterns = [
             url(r'^example/?$', views.ExampleListView.as_view()),
         ]
-        generator = OpenAPISchemaGenerator(patterns=patterns)
+        generator = SchemaGenerator(patterns=patterns)
 
         request = create_request('/')
         schema = generator.get_schema(request=request)
